+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [0.4.9] - Unreleased
 ### Added
 * Support for yaml merge
+* MwsNccBinding databricks resource
+* MwsNetworkConnectivityConfig databricks resource
 ### Fixed
 * DLT Debug mode when source is streaming and node is not
 * DataFrame type propagation when models are used as inputs to other models